trialcode/requirements.txt
torch>=1.9.0
grpcio>=1.51.0
grpcio-tools>=1.51.0
pyyaml>=6.0
scikit-learn>=1.1.0
pandas>=1.3.0
numpy>=1.21.0
